requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
pytest>=7.0.0
psutil>=5.8.0